# openapi.yaml
openapi: 3.1.0
info:
  title: Contacts api's docs
  description : This is the documentation of organization management system 
  version : 2.0.0

servers:
 - url: http://127.0.0.1:8000/api/v1
   description: Production
 - url: http://127.0.0.1:8000/api/v2
   description: Dev

tags:
  - name: Employee
    description: Routes that are specific to employees resource
  # - name: General
  #   description: These are the routes that can be accessed by any user
  - name: Auth
    description: These are the routes employees hit when dealing with auth features 

paths:
  /auth/employee/{emp_id}/password:
    post:
      summary: This route will be used to create a new password
      tags:
        - Auth
      parameters:
        - $ref: "#/components/parameters/emp_id"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                new_pass:
                  $ref: "#/components/schemas/password"
                old_pass:
                  $ref: "#/components/schemas/password"
              required:
                - new_pass 
                - old_pass

      responses:
        204:
          description: Password updated successfully 
        401:
          $ref: "#/components/responses/401Unauthorized"
  /auth/login:
    post:
      summary: This is the route we will hit to login and get token
      tags:
        - Auth
      requestBody:
        $ref: "#/components/requestBodies/login_body"
      responses:
        200:
          description: A login token 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/emp_info"
        401:
          $ref: "#/components/responses/401Unauthorized"
        500:
          $ref: "#/components/responses/500InternalServerError"

  /employee:
    get:
      security:
        - contacts_auth: []
      tags:
        - Employee
      parameters:
        - name: query_username
          required: true
          in: query
          schema:
            type: string

      summary: Returns a list of users according to search query.
      description: Returns an array of username and id that match the search query
      responses:
        200:
          description: A json array of usernames and user ids
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    examples:
                      - True
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/emp_info"
        401:
          $ref: "#/components/responses/401Unauthorized"
        500:
          $ref: "#/components/responses/500InternalServerError"
    
    post:
      security:
        - contacts_auth: []
      tags:
        - Employee
      summary: Creates a user in the system 
      description: Endpoint creates a user in the system, this endpoint can only be hit by admin
      requestBody:
        content: 
          application/json:
            schema:
              $ref: "#/components/schemas/newEmployee"
      responses:
        201:
          $ref: "#/components/responses/201EmployeeCreated"
        400:
          $ref: "#/components/responses/400FieldValidationFailed"
        401:
          $ref: "#/components/responses/401Unauthorized"
        403:
          $ref: "#/components/responses/403Forbidden"
        409:
          $ref: "#/components/responses/409EmailConflict"
        500:
          $ref: "#/components/responses/500InternalServerError"
  /employee/{emp_id}:
    delete:
      security:
        - contacts_auth: []
      tags:
        - Employee
      parameters:
        - $ref: "#/components/parameters/emp_id"

      summary: Removes an employee from the database
      description: Endpoint removes an employee from the database can only be invoked by admin will shift all the reporting employees to the person next in hierarchy

      responses:
        204:
          description: A simple response that tells if deletion is successful
        400:
          $ref: "#/components/responses/400BadRequest"
        401:
          $ref: "#/components/responses/401Unauthorized"
        403:
          $ref: "#/components/responses/403Forbidden"
        404:
          $ref: "#/components/responses/404NotFound"
        405:
          $ref: "#/components/responses/405NotAllowed"
        500:
          $ref: "#/components/responses/500InternalServerError"

  /employee/profile:
    get:
      security:
        - contacts_auth: []
      tags:
        - Employee
      summary: Gets employee's own profile info 
      description: Endpoint fetches the profile info of an employee can only be accessed by user themself
      responses:
        200:
          $ref: "#/components/responses/200ProfileInfo"
        401:
          $ref: "#/components/responses/401Unauthorized"
        404:
          $ref: "#/components/responses/404NotFound"
        500:
          $ref: "#/components/responses/500InternalServerError"
        
  /employee/request:
    get:
      tags: 
        - Employee
      summary: Returns a list of requests created by employee 
      parameters:
        - $ref: '#/components/parameters/request_status'
      security:
        - contacts_auth: []
      responses:
        200:
          $ref: "#/components/responses/200requests"
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: "#/components/responses/401Unauthorized"
        403:
          $ref: "#/components/responses/403Forbidden"
        404:
          $ref: "#/components/responses/404NotFound"
        500:
          $ref: "#/components/responses/500InternalServerError"

    post:
      tags:
        - Employee
      summary: Creates a new request in the db
      requestBody:
        $ref: "#/components/requestBodies/create_request_body"
      security:
        - contacts_auth: []
      responses:
        
        201:
          $ref: "#/components/responses/201RequestCreated"
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: "#/components/responses/401Unauthorized"
        403:
          $ref: "#/components/responses/403Forbidden"
        500:
          $ref: "#/components/responses/500InternalServerError"
  
  /employee/{emp_id}/reports_to:
    get:
      security:
        - contacts_auth: []
      tags:
        - Employee
      summary: Returns the employee to whom an employee with id equal to emp_id reports 
      description: This endpoint behaves differently based on the current logged in user for admin it only return the user id of the person reporting to and for worker and hr it returns the name and contact info of the employee 
      parameters:
        - $ref: "#/components/parameters/emp_id"
      responses:
        200:
          description: Basic contact info of the user  
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    examples:
                      - True
                  data:
                    $ref: "#/components/schemas/relation_info"
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: "#/components/responses/401Unauthorized"
        500:
          $ref: "#/components/responses/500InternalServerError"
    
    put:
      security:
        - contacts_auth: []
      tags:
        - Employee
      summary: This endpoint will change who a user reports to can only be hit by an admin
      description: If user does not report to anyone reports_to field will have value of 0  
      parameters:
        - $ref: "#/components/parameters/emp_id"
      requestBody:
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports_to:
                    $ref: "#/components/schemas/emp_id"
                required:
                  - reports_to
      responses:
        204:
          description: Simple 204 successful request fulfilment but not returning any response
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: "#/components/responses/401Unauthorized"
        403:
          $ref: "#/components/responses/403Forbidden"
        500:
          $ref: "#/components/responses/500InternalServerError"
  /employee/{emp_id}/reported_by:
    get:
      security:
        - contacts_auth: []
      tags:
        - Employee 
      summary: Returns an array of employees that report to employee with id is equal to emp_id
      parameters:
        - $ref: "#/components/parameters/emp_id"
      responses:
        200:
          description: Basic contact info of the users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    examples:
                      - True
                  data:
                    type: array 
                    items:
                      $ref: "#/components/schemas/emp_info"
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: "#/components/responses/401Unauthorized"
        500:
          $ref: "#/components/responses/500InternalServerError"


  /admin/request:
    get:
      tags:
        - Admin
      summary: Returns a list of requests  
      parameters:
      - $ref: "#/components/parameters/request_status"
      security:
        - contacts_auth: []
      responses:
        200:
         $ref: "#/components/responses/200requests"
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: "#/components/responses/401Unauthorized"
        403:
          $ref: "#/components/responses/403Forbidden"
        500:
          $ref: "#/components/responses/500InternalServerError"

  /admin/request/{req_id}:
    patch:
      tags:
        - Admin
      summary: Commits an employee's request into the database hence updating the user 
      parameters:
        - $ref: "#/components/parameters/req_id"
      security:
        - contacts_auth: []
      responses:
        200:
          $ref: "#/components/responses/200requests"
        400:
          $ref: "#/components/responses/400BadRequest"
        401:
          $ref: "#/components/responses/401Unauthorized"
        403:
          $ref: "#/components/responses/403Forbidden"
        500:
          $ref: "#/components/responses/500InternalServerError"

      
          
  /hr/request:
    get:
      tags:
        - Hr
      summary: Returns a list of requests that are assigned to logged in Hr  
      parameters:
        - $ref: "#/components/parameters/request_status"
      security:
        - contacts_auth: []
      responses:
        200:
          $ref: "#/components/responses/200requests"
        400:
          $ref: "#/components/responses/400BadRequest "
        401:
          $ref: "#/components/responses/401Unauthorized "
        403:
          $ref: "#/components/responses/403Forbidden"
        500:
          $ref: "#/components/responses/500InternalServerError"
      
  /hr/request/{req_id}:
    patch:
      tags:
        - Hr
      summary: Updates the request accordingly and returns a list of requests  
      parameters:
      - $ref: "#/components/parameters/req_id"

      requestBody:
        $ref: "#/components/requestBodies/HrUpdateReqStatus"

      security:
        - contacts_auth: []
      responses:
        200:
          $ref: "#/components/responses/200requests"
        400:
          $ref: "#/components/responses/400BadRequest"
        401:
          $ref: "#/components/responses/401Unauthorized"
        403:
          $ref: "#/components/responses/403Forbidden"
        500:
          $ref: "#/components/responses/500InternalServerError"

      
components:
  securitySchemes:
    # bearerAuth: # arbitrary name for the security scheme
    #   type: http
    #   scheme: bearer
    #   bearerFormat: JWT
  
    contacts_auth:
      type: oauth2
      flows:
        password:
          tokenUrl: /api/v1/auth/login

  schemas:
    request: 
      type: object
      properties: 
        request_id:
          type: integer
        created_by:
          $ref: "#/components/schemas/name"
        updated_info:
          $ref: "#/components/schemas/updatedInfo"
        assigned_hr:
          $ref: "#/components/schemas/name"
        created_at:
          $ref: "#/components/schemas/timeStamp"
        update_committed_at:
          $ref: "#/components/schemas/timeStamp"
        request_status: 
          type: string
          enum: ["hr_assigned","committed","rejected","approved_by_hr"]
        remark: 
          oneOf:
           - type: string 
           - type: "null" 
      required:
        - reqId
        - created_by
        - updated_info
        - assigned_hr
        - created_at
        - update_committed_at
        - request_status
    updatedInfo:
      type: object
      properties:
        name: 
          $ref: "#/components/schemas/name"
        phone:
          $ref: "#/components/schemas/phone"
        email:
          $ref: "#/components/schemas/email"
        address: 
          $ref: '#/components/schemas/address'
      
      required:
        - name
        - phone
        - email
        - address
    
    address:
      type: object
      properties:
        street:
          type: string
        postal_code:
          type: integer
          examples:
            - 243001
        city:
          type: string
        state:
          type: string
        country:
          type: string
          
      required:
        - street
        - postal_code
        - city
        - state
        - country
    emp_id:
      name: employeeId
      type: integer
      examples:
        - 4
      description: Employee id
    
    timeStamp:
      name: timeStamp
      type: integer
      examples:
        - 1727347857
      description: Unix epoch timestamp in seconds  

    emp_info:
      type: object
      properties:
        name:
          $ref: "#/components/schemas/name"
        emp_id:
          $ref: "#/components/schemas/emp_id"
        email:
          $ref: "#/components/schemas/email"
        phone:
          $ref: "#/components/schemas/phone"
      required:
        - id
        - name
        - email
        - phone
      examples:
      - id: 9
        name: sampleName
        email: test@test.com
        phone: 9999999999
    
    relation_info:
      type: object
      properties: 
        reports_to:
            $ref: "#/components/schemas/emp_info"
        employee: 
            $ref: "#/components/schemas/emp_info"

    relation:
      type: object
      properties:
        employee: 
          type: integer
        reports_to:
          type: integer

    newEmployee:
      type: object
      properties:
        user_type:
          $ref: "#/components/schemas/user_type"
        name:
          $ref: "#/components/schemas/name"
        phone:
          $ref: "#/components/schemas/phone"
        email: 
          $ref: "#/components/schemas/email"
        address: 
          $ref: '#/components/schemas/address'
        password:
          $ref: "#/components/schemas/password"
        reports_to:
          type: integer
          description: userId of employee which this employee reports to
      required:
        - name
        - userType
        - phone
        - email
        - address
        - reports_to
    user_type: 
      enum: ["hr","admin","worker"]

    name: 
      type: string
      examples:
        - Test user
    email:
      type: string
      examples:
        - test@user.com
    phone:
        type: integer
        examples: 
        - 1234567890
    password:
      description: There must be 1 int 1 uppercase 1 lowercase and 1 special character with min length = 8 
      type: string
      examples: 
        - Test@123

  requestBodies:
    login_body:
      required: true
      content:
        multipart/form-data:
          schema:
            type: object
            properties:
              password: 
                type: string
              username:
                type: integer
            
    create_request_body:
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              name:
                  $ref: "#/components/schemas/name" 
              phone: 
                  $ref: "#/components/schemas/phone" 
              email:
                  $ref: "#/components/schemas/email" 
              address: 
                  $ref: "#/components/schemas/address" 

    HrUpdateReqStatus:
      required: true
      content: 
        application/json:
          schema:
            type: object
            properties: 
              request_status:
                type: string
                examples:
                  - approved_by_hr
              remark:
                type: string
                examples:
                  - this is a remark
            required:
              - request_status
              - remark 
      
  parameters:
    request_status:
      name: status
      in: query
      required: false 
      description: status of the request can be a csv string of committed,rejected, approved_by_hr, hr_assigned 
      schema:
        type: string

    authorization:
      name: authorization
      in: header
      required: true
      description: This is the auth token containing login info and userId
      schema:
        type: string

    req_id:
      name: req_id
      in: path
      required: true
      schema:
        type: integer

    emp_id:
      name: emp_id
      in: path
      required: true
      schema:
        type: integer 
   
  responses:
    200Success:
      description: A simple 200 success response
      content:
        application/json: 
          schema:
            type: object
            properties:
              success: 
                type: boolean
              message:
                type: string
            required:
              - success
              - message
    201EmployeeCreated:
      description: Employee created response
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
              emp_id: 
                type: integer
            required:
              - success
              - emp_id

    201RequestCreated:
      description: Request created response
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
              request_id: 
                type: integer
            required:
              - success
              - request_id 

    200ProfileInfo:
      description: A simple response containing address, employee Id, name, phone number and email 
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
              profile_info:
                type: object
                properties:
                  name:
                    $ref: "#/components/schemas/name"
                  email:
                    $ref: "#/components/schemas/email"
                  phone:
                    $ref: "#/components/schemas/phone"
                  address:
                    $ref: "#/components/schemas/address"
                  emp_id:
                    $ref: "#/components/schemas/emp_id"
    200requests:
      description: All requests that fall in the query catagories 
      content:
        application/json: 
          schema:
            type: object
            properties:
              success: 
                type: boolean
              data:
                type: array
                items:
                  $ref: "#/components/schemas/request"

    400BadRequest:
      description: "Bad request"
      content:
        application/json:
          schema:
            type: object
            properties: 
              success: 
                type: boolean
                examples:
                  - false
              error:
                type: string
                examples:
                  - Bad request 

    400FieldValidationFailed:
      description: "Some or all of the fields provided in the request are ivalid"
      content:
        application/json:
          schema:
            type: object
            properties:
              success: 
                type: boolean
                examples:
                  - false
              error:
                type: String
                examples:
                  - Invalid value for email

    401Unauthorized:
      description: "Not authorized"
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                examples:
                  - false
              error:
                type: string
                examples:
                  - Not logged in              

    403Forbidden:
      description: "Accessing restricted route"
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                examples:
                  - false
              error:
                type: string
                examples:
                  - Restricted route

    404NotFound:
      description: "Resource requested is not found"
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                examples:
                  - false
              error:
                type: string
                examples:
                  - Requested resource not found

    405NotAllowed:
      description: "Method requested on the resource not allowed"
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                examples:
                  - false
              error:
                type: string
                examples:
                  - Requested resource can not be deleted 

    409EmailConflict:
      description: "User already exists"
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                examples:
                  - false
              message:
                type: string
                examples:
                  - User with email test@test.com already exists


    500InternalServerError:
      description: "Internal server error"
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                examples:
                  - false
              error:
                type: string
                examples:
                  - There was some error while processing the request 